Use CSS transitions
Prefer CSS transition because that animation is executed the different
thread from main thread. Current hardwares have more cpus and then
parallel execution will be smoother in the future.
Also gsap is not compatible with experimentalHoisting option of Parcel.
We should completely remove that dependency and minimize output
javascript file.